Move all files into `app` directory
It's better than having some files at the root
