+++ Dockerfile
@@ -7,7 +7,7 @@ WORKDIR /app
 RUN poetry config virtualenvs.in-project true
 RUN poetry install --no-dev --no-root
 
-ENTRYPOINT ["poetry", "run", "gunicorn", "-k", "uvicorn.workers.UvicornWorker", "--access-logfile", "-", "--bind", "0.0.0.0:80", "--timeout", "60", "main:app"]
+ENTRYPOINT ["poetry", "run", "gunicorn", "-k", "uvicorn.workers.UvicornWorker", "--access-logfile", "-", "--bind", "0.0.0.0:80", "--timeout", "60", "app:app"]
 
 EXPOSE 80
